More enemy variaty

+++ Game_for_school/Assets/Scripts/Enemy_stats.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public class enemy_stats : MonoBehaviour
 {
-    public int health;
-    public int maxHealth;
-    public int damage;
-    public int stamina;
-    public int maxStamina;
+    public float health;
+    public float maxHealth;
+    public float damage;
+    public float stamina;
+    public float maxStamina;
     public bool isMiniboss;
     public bool isBoss;
